# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: API de Notificação
  description: Notificação de pedidos e orçamentos
host: localhost:8010
basePath: /api/v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /notificacoes/pedidos/solicitacao:
    post:
      summary: Notificação de solicitação
      description: 'Notifica o atacadista com os dados do pedido.'
      operationId: solicitarPedido
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SolicitacaoPedidoBodyPost'
      tags:
        - Notificacao            
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SucessResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
      x-swagger-router-controller: "Notificacao"
  /notificacoes/pedidos/orcamento:
    post:
      summary: Notificação de Orçamento
      description: 'Notifica o logista com os dados do orçamento'
      operationId: enviarOrcamento
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/OrcamentoBodyPost'
      tags:
        - Notificacao              
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SucessResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
      x-swagger-router-controller: "Notificacao"
  /notificacoes/pedidos/orcamento/aprova:
    post:
      summary: Notificação de confirmação
      description: 'Notifica a confirmação do pedido'
      operationId: aprovarPedido
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AprovaBodyPost'
      tags:
        - Notificacao    
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SucessResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
      x-swagger-router-controller: "Notificacao"            
  /notificacoes/pedidos/orcamento/reprova:
    post:
      summary: Notificação de cancelamento
      description: 'Notifica o cancelamento do pedido'
      operationId: reprovarPedido
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ReprovaBodyPost'
      tags:
        - Notificacao
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SucessResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
      x-swagger-router-controller: "Notificacao"            
  /notificacoes/pedidos/situacao:
    post:
      summary: Notificação de Situação  
      description: 'Notifica a situação do pedido'
      operationId: verificarSituacaoPedido
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SituacaoBodyPost'
      tags:
        - Notificacao            
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SucessResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"       
      x-swagger-router-controller: "Notificacao"            
definitions:
  SucessResponse:
    required:
      - message
    properties:
      message:
        type: string   
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SolicitacaoPedidoBodyPost:
    type: object
    properties:
      codigo:
        type: integer
        description: Código identificador do pedido
      quantidade:
        type: integer
        description: Quantidade do pedido
      observacao:
        type: string
        description: Observações do pedido
  OrcamentoBodyPost:
    type: object
    properties:
      numero:
        type: integer
        description: Código identificador do orçamento
      entrega:
        type: string
        format: date
        description: Data de entrega do pedido
  AprovaBodyPost:
    type: object
    properties:
      numero:
        type: integer
        description: Código identificador do orçamento
      entrega:
        type: string
        format: date
        description: Data de entrega do pedido      
  ReprovaBodyPost:
    type: object
    properties:
      numero:
        type: integer
        description: Código identificador do orçamento
      entrega:
        type: string
        format: date
        description: Data de entrega do pedido 
  SituacaoBodyPost:
    type: object
    properties:
      numero:
        type: integer
        description: Código identificador do orçamento
      situacao:
        type: string
        description: Situação em que o pedido se encontra          